==> etc/10.10-descriptive.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
from pathlib import Path
if "../src" not in sys.path:
    sys.path.append("../src")
from tqdm.auto import tqdm
import numpy as np

DYNSPANS="tier1,先秦,漢,魏晉南北,唐五代十國,宋元,明,清".split(",")

import json
from collections import Counter
from itertools import chain
def flatten(x):
    if isinstance(x, str):
        return x
    elif isinstance(x, list):
        return "".join(x)
    else:
        assert False, ""

records = []
for dyn_x in DYNSPANS:
    with open(f"../data/corpus/data/dynasty_split/{dyn_x}.jsonl", "r") as fin:
        dyn_counter = {}
        ntext = 0
        for ln in tqdm(fin, desc=dyn_x):
            ntext += 1
            text = json.loads(ln)        
            char_counter = Counter(chain.from_iterable(flatten(x["c"]) for x in text["text"]))
            dyn_counter.update(char_counter)
        records.append((dyn_x, ntext, len(dyn_counter), sum(dyn_counter.values())))
        

import pandas as pd
dynfreq = pd.DataFrame(records, columns="dyn ntext ntype ntoken".split())

dynfreq

dynfreq.ntext.sum()-120
