==> etc/11.01-author_mapping.ipynb <==
import sys, os
from pathlib import Path
if "../src" not in sys.path:
    sys.path.append("../src")
    
import re
import pandas as pd
from collections import defaultdict

save_mode = True # if True, save generated data except for wiki_retrieved_data
post_edit_mode = True # if True, add author time from notes

from dietcoke.get_author import get_all_authors
authors_tier1, authors_tier12, authors_uni = get_all_authors(save_names_clean=save_mode)

## print
from dietcoke.get_author import PAT_ANONYM
cnt_tier1 = sum([not re.match(PAT_ANONYM, n) for n in authors_tier1])
cnt_tier12 = sum([not re.match(PAT_ANONYM, n) for n in authors_tier12])

print('---\nCount of tier1:', cnt_tier1, '/', len(authors_tier1))
print('Coverage of tier1:', round(cnt_tier1 / len(authors_tier1), 2))

print('---\nCount of tier12:', cnt_tier12, '/', len(authors_tier12))
print('Coverage of tier12:', round(cnt_tier12 / len(authors_tier12), 2))

# from dietcoke.get_author import get_wiki_data
# get_wiki_data()

---

from dietcoke.get_author import get_wiki_author_time
get_wiki_author_time(save_wiki_author_time=save_mode)

from dietcoke.author import get_author_time_map
author_time = get_author_time_map()
print(len(author_time))

sorted(author_time.items(), key=lambda x: x[1])[:10]

from dietcoke import Corpus, dynaspan_lst, Author

for dynaspan in dynaspan_lst:
    print('-----\n', dynaspan)
    corpus = Corpus(dynaspan)
    corpus.read_corpus()

    c = 0
    for i, line in enumerate(corpus.corpus):
        rep_year = Author(line.author).rep_year
        if rep_year != -9999:
            c += 1
    print(c, '/', i)

from dietcoke.author import get_author_time_map
group_authors = defaultdict(list)
for author, time in get_author_time_map(merge_data=False).items():
    group_authors[str(time)].append(author)

for life, author_lst in group_authors.items():
    if len(author_lst) >= 2:
        print(life, author_lst)

df = pd.DataFrame(list(group_authors.items()))
df.columns = ('life', 'author_lst')
df['mid_year'] = [Author(n[-1]).rep_year for n in df['author_lst']]
df.sort_values('mid_year')

if save_mode:
    df.to_csv('../data/author_time/author_timeline.csv')

df

# 莊述祖
# 王树枏/王树枬

---

from tqdm.auto import tqdm
from dietcoke import corpus_lst, Author
import pandas as pd

authors = []
profiles = []
for corpus in tqdm(corpus_lst()):
    corpus.read_corpus()
    dynaspan = corpus.dynaspan
    for line in corpus.corpus:
        line_authors = Author(line.author).name_norm
        for profile in Author(line.author).author_profile:
            profile['urn'] = line.urn
            profile['title'] = line.obj['title']
            profile['dynaspan'] = dynaspan
            profiles.append(profile)
            line_authors.remove(profile['name'])
        authors += [{'name': name,
        'urn': line.urn, 'title': line.obj['title'], 'dynaspan': dynaspan} for name in line_authors]

profile_df = pd.DataFrame(profiles).sort_values(['birth_year', 'name'])
author_df = pd.DataFrame(authors).sort_values('name')
profile_df = pd.merge(profile_df, author_df, how='outer')

if save_mode:
    profile_df.to_csv('author_profile.csv')
