==> etc/40.01-dynasty-emb.ipynb <==
from pathlib import Path
import json
import numpy as np
from tqdm.auto import tqdm
from gensim import corpora
from gensim.models import LsiModel
from gensim.models import KeyedVectors

def extract_text(text_path):
    fin = open(text_path, "r", encoding="UTF-8")
    json_objs = []
    for ln in fin:
        json_data = json.loads(ln)
        json_objs.append(json_data)
    fin.close()
    
    text = []
    for j_idx, jobj in enumerate(tqdm(json_objs)):
        for text_idx, text_obj in enumerate(jobj["text"]):            
            if isinstance(text_obj["c"], list):
                text.append(list(text_obj["t"]) + 
                            list("\n".join(text_obj["c"])))
            else:
                text.append(list(text_obj["t"]) + 
                            list(text_obj["c"]))
    return text

text_path_list = list(Path("../corpus/dynasty_split/").glob("*.jsonl"))
text_path_list

dyn0 = extract_text("../corpus/dynasty_split/先秦.jsonl")
dyn1 = extract_text("../corpus/dynasty_split/漢.jsonl")
dyn2 = extract_text('../corpus/dynasty_split/魏晉南北.jsonl')

len(dyn0), len(dyn1), len(dyn2)

## Basic statistics

print("dyn0 char count", sum(len(x) for x in dyn0))
print("dyn1 char count", sum(len(x) for x in dyn1))
print("dyn2 char count", sum(len(x) for x in dyn2))

def make_lsa_vec(dyn_data):
    vocab = corpora.Dictionary(dyn_data)
    vocab.filter_extremes(no_below=5)    
    docmat = [vocab.doc2bow(x) for x in dyn_data]
    lsa = LsiModel(docmat)
    lsavec = lsa.projection.u * lsa.projection.s
    kv = KeyedVectors(200)    
    kv.add_vectors(list(vocab.token2id.keys()), lsavec)
    print("vocab: ", len(kv.index_to_key))
    return kv

print("Running dyn0"); kv0 = make_lsa_vec(dyn0)
print("Running dyn1"); kv1 = make_lsa_vec(dyn1)
print("Running dyn2"); kv2 = make_lsa_vec(dyn2)

kv_list = [
    {"kv": kv0, "note": "dynasty_split/先秦.jsonl"},
    {"kv": kv1, "note": "dynasty_split/漢.jsonl"},
    {"kv": kv2, "note": "dynasty_split/魏晉南北.jsonl"}
]

import joblib
kv_list_path = joblib.dump(kv_list, "../data/dyn_kvlist_3dyns.joblib")[0]
!sha1sum $kv_list_path

list(set(kv0.index_to_key) & set(kv1.index_to_key) & set(kv2.index_to_key))[:10]

kv0.most_similar("雪")

kv1.most_similar("雪")

kv2.most_similar("雪")
